Forgot to commit last time I made changes. It looks like the last time I should have commited just had changes requested by the style analysis of my code by VS.
This time I made minor updates to logic and added a new feature. IsAffected now compares the strings instead of their lengths. You can also request help by clicking on the link in the lower right of app. The help link takes you to the MSDN docs for regex in .NET 4.5 in your default browswer.